#  storage/ api-docs/ api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api Project",
    "description": "Demo my Project ",
    "contact": {
      "email": "admin@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/category": {
      "get": {
        "tags": [
          "Category"
        ],
        "operationId": "327644f4d836240fa297b2822ab6ca92",
        "responses": {
          "200": {
            "description": "List Categories."
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "operationId": "de40ad04180a48262a05cf249f785ab3",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "name",
                  "image",
                  "description"
                ],
                "properties": {
                  "image": {
                    "type": "file"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add Category."
          }
        }
      }
    },
    "/api/category/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "operationId": "b5e0594038139a2979e671987221c993",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ідентифікатор категорії",
            "required": true,
            "schema": {
              "type": "number",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List Categories."
          },
          "404": {
            "description": "Категорії не знайдено"
          },
          "401": {
            "description": "Не авторизований"
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "operationId": "d2437e8e0e200592efe7305af1c302d7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ідентифікатор категорії",
            "required": true,
            "schema": {
              "type": "number",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успішне видалення категорії"
          },
          "404": {
            "description": "Категорії не знайдено"
          },
          "401": {
            "description": "Не авторизований"
          }
        }
      }
    },
    "/api/category/edit/{id}": {
      "post": {
        "tags": [
          "Category"
        ],
        "operationId": "94a8432736ed864d323e7cb413519b8d",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ідентифікатор категорії",
            "required": true,
            "schema": {
              "type": "number",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "name"
                ],
                "properties": {
                  "image": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add Category."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  }
}